fix: refactor higher leverl interfaces into io
