+++ protocol/src/io.rs
@@ -1,0 +1,381 @@
+// SPDX-License-Identifier: CC0-1.0
+
+//! High-level interfaces for establishing and using BIP324 encrypted
+//! connections over Read/Write transports.
+
+use core::fmt;
+
+#[cfg(feature = "alloc")]
+use alloc::vec;
+#[cfg(feature = "alloc")]
+use alloc::vec::Vec;
+
+use bitcoin::Network;
+
+// Default to the futures-rs traits, but can overwrite with more specific
+// tokio implementations for easier caller integration.
+#[cfg(all(feature = "futures", not(feature = "tokio")))]
+use futures::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
+#[cfg(feature = "tokio")]
+use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
+
+use crate::{
+    Error, Handshake, PacketReader, PacketType, PacketWriter, Payload, Role,
+    NUM_ELLIGATOR_SWIFT_BYTES, NUM_GARBAGE_TERMINTOR_BYTES, NUM_INITIAL_HANDSHAKE_BUFFER_BYTES,
+    VERSION_CONTENT,
+};
+
+/// High level error type for the protocol interface.
+#[cfg(feature = "std")]
+#[derive(Debug)]
+pub enum ProtocolError {
+    /// Wrap all IO errors with suggestion for next step on failure.
+    Io(std::io::Error, ProtocolFailureSuggestion),
+    /// Internal protocol specific errors.
+    Internal(Error),
+}
+
+/// Suggest to caller next step on protocol failure.
+#[cfg(feature = "std")]
+#[derive(Debug)]
+pub enum ProtocolFailureSuggestion {
+    /// Caller could attempt to retry the connection with protocol V1 if desired.
+    RetryV1,
+    /// Caller should not attempt to retry connection.
+    Abort,
+}
+
+#[cfg(feature = "std")]
+impl From<std::io::Error> for ProtocolError {
+    fn from(error: std::io::Error) -> Self {
+        // Detect IO errors which possibly mean the remote doesn't understand
+        // the V2 protocol and immediately closed the connection.
+        let suggestion = match error.kind() {
+            // The remote force closed the connection.
+            std::io::ErrorKind::ConnectionReset
+            // A more general error than ConnectionReset, but could be caused
+            // by the remote closing the connection.
+            | std::io::ErrorKind::ConnectionAborted
+            // End of file read errors can occur if the remote closes the connection,
+            // but the local system reads due to timing issues.
+            | std::io::ErrorKind::UnexpectedEof => ProtocolFailureSuggestion::RetryV1,
+            _ => ProtocolFailureSuggestion::Abort,
+        };
+
+        ProtocolError::Io(error, suggestion)
+    }
+}
+
+#[cfg(feature = "std")]
+impl From<Error> for ProtocolError {
+    fn from(error: Error) -> Self {
+        ProtocolError::Internal(error)
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for ProtocolError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        match self {
+            ProtocolError::Io(e, _) => Some(e),
+            ProtocolError::Internal(e) => Some(e),
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+impl fmt::Display for ProtocolError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ProtocolError::Io(e, suggestion) => {
+                write!(
+                    f,
+                    "IO error: {}. Suggestion: {}.",
+                    e,
+                    match suggestion {
+                        ProtocolFailureSuggestion::RetryV1 => "Retry with V1 protocol",
+                        ProtocolFailureSuggestion::Abort => "Abort, do not retry",
+                    }
+                )
+            }
+            ProtocolError::Internal(e) => write!(f, "Internal error: {e}."),
+        }
+    }
+}
+
+/// A protocol session with handshake and send/receive packet management.
+#[cfg(any(feature = "futures", feature = "tokio"))]
+pub struct AsyncProtocol {
+    reader: AsyncProtocolReader,
+    writer: AsyncProtocolWriter,
+}
+
+#[cfg(any(feature = "futures", feature = "tokio"))]
+impl AsyncProtocol {
+    /// New protocol session which completes the initial handshake and returns a handler.
+    ///
+    /// # Arguments
+    ///
+    /// * `network` - Network which both parties are operating on.
+    /// * `role` - Role in handshake, initiator or responder.
+    /// * `garbage` - Optional garbage bytes to send in handshake.
+    /// * `decoys` - Optional decoy packet contents bytes to send in handshake.
+    /// * `reader` - Asynchronous buffer to read packets sent by peer.
+    /// * `writer` - Asynchronous buffer to write packets to peer.
+    ///
+    /// # Returns
+    ///
+    /// A `Result` containing:
+    ///   * `Ok(AsyncProtocol)`: An initialized protocol handler.
+    ///   * `Err(ProtocolError)`: An error that occurred during the handshake.
+    ///
+    /// # Errors
+    ///
+    /// * `Io` - Includes a flag for if the remote probably only understands the V1 protocol.
+    pub async fn new<'a, R, W>(
+        network: Network,
+        role: Role,
+        garbage: Option<&'a [u8]>,
+        decoys: Option<&'a [&'a [u8]]>,
+        reader: &mut R,
+        writer: &mut W,
+    ) -> Result<Self, ProtocolError>
+    where
+        R: AsyncRead + Unpin + Send,
+        W: AsyncWrite + Unpin + Send,
+    {
+        let garbage_len = match garbage {
+            Some(slice) => slice.len(),
+            None => 0,
+        };
+        let mut ellswift_buffer = vec![0u8; NUM_ELLIGATOR_SWIFT_BYTES + garbage_len];
+        let mut handshake = Handshake::new(network, role, garbage, &mut ellswift_buffer)?;
+
+        // Send initial key to remote.
+        writer.write_all(&ellswift_buffer).await?;
+        writer.flush().await?;
+
+        // Read remote's initial key.
+        let mut remote_ellswift_buffer = [0u8; 64];
+        reader.read_exact(&mut remote_ellswift_buffer).await?;
+
+        let num_version_packet_bytes = PacketWriter::required_packet_allocation(&VERSION_CONTENT);
+        let num_decoy_packets_bytes: usize = match decoys {
+            Some(decoys) => decoys
+                .iter()
+                .map(|decoy| PacketWriter::required_packet_allocation(decoy))
+                .sum(),
+            None => 0,
+        };
+
+        // Complete materials and send terminator to remote.
+        // Not exposing decoy packets yet.
+        let mut terminator_and_version_buffer =
+            vec![
+                0u8;
+                NUM_GARBAGE_TERMINTOR_BYTES + num_version_packet_bytes + num_decoy_packets_bytes
+            ];
+        handshake.complete_materials(
+            remote_ellswift_buffer,
+            &mut terminator_and_version_buffer,
+            decoys,
+        )?;
+        writer.write_all(&terminator_and_version_buffer).await?;
+        writer.flush().await?;
+
+        // Receive and authenticate remote garbage and version.
+        // Keep pulling bytes from the buffer until the garbage is flushed.
+        let mut remote_garbage_and_version_buffer =
+            Vec::with_capacity(NUM_INITIAL_HANDSHAKE_BUFFER_BYTES);
+        loop {
+            let mut temp_buffer = [0u8; NUM_INITIAL_HANDSHAKE_BUFFER_BYTES];
+            match reader.read(&mut temp_buffer).await {
+                // No data available right now, continue.
+                Ok(0) => {
+                    continue;
+                }
+                Ok(bytes_read) => {
+                    remote_garbage_and_version_buffer.extend_from_slice(&temp_buffer[..bytes_read]);
+
+                    match handshake
+                        .authenticate_garbage_and_version(&remote_garbage_and_version_buffer)
+                    {
+                        Ok(()) => break,
+                        // Not enough data, continue reading.
+                        Err(Error::CiphertextTooSmall) => continue,
+                        Err(e) => return Err(ProtocolError::Internal(e)),
+                    }
+                }
+                Err(e) => match e.kind() {
+                    // No data available or interrupted, retry.
+                    std::io::ErrorKind::WouldBlock | std::io::ErrorKind::Interrupted => {
+                        continue;
+                    }
+                    _ => return Err(ProtocolError::Io(e, ProtocolFailureSuggestion::Abort)),
+                },
+            }
+        }
+
+        let packet_handler = handshake.finalize()?;
+        let (packet_reader, packet_writer) = packet_handler.into_split();
+
+        Ok(Self {
+            reader: AsyncProtocolReader {
+                packet_reader,
+                state: DecryptState::init_reading_length(),
+            },
+            writer: AsyncProtocolWriter { packet_writer },
+        })
+    }
+
+    /// Read reference for packet reading operations.
+    pub fn reader(&mut self) -> &mut AsyncProtocolReader {
+        &mut self.reader
+    }
+
+    /// Write reference for packet writing operations.
+    pub fn writer(&mut self) -> &mut AsyncProtocolWriter {
+        &mut self.writer
+    }
+
+    /// Split the protocol into a separate reader and writer.
+    pub fn into_split(self) -> (AsyncProtocolReader, AsyncProtocolWriter) {
+        (self.reader, self.writer)
+    }
+}
+
+/// State machine of an asynchronous packet read.
+#[cfg(any(feature = "futures", feature = "tokio"))]
+#[derive(Debug)]
+enum DecryptState {
+    ReadingLength {
+        length_bytes: [u8; 3],
+        bytes_read: usize,
+    },
+    ReadingPayload {
+        packet_bytes: Vec<u8>,
+        bytes_read: usize,
+    },
+}
+
+#[cfg(any(feature = "futures", feature = "tokio"))]
+impl DecryptState {
+    /// Transition state to reading the length bytes.
+    fn init_reading_length() -> Self {
+        DecryptState::ReadingLength {
+            length_bytes: [0u8; 3],
+            bytes_read: 0,
+        }
+    }
+
+    /// Transition state to reading payload bytes.
+    fn init_reading_payload(packet_bytes_len: usize) -> Self {
+        DecryptState::ReadingPayload {
+            packet_bytes: vec![0u8; packet_bytes_len],
+            bytes_read: 0,
+        }
+    }
+}
+
+/// Manages an async buffer to automatically decrypt contents of received packets.
+#[cfg(any(feature = "futures", feature = "tokio"))]
+pub struct AsyncProtocolReader {
+    packet_reader: PacketReader,
+    state: DecryptState,
+}
+
+#[cfg(any(feature = "futures", feature = "tokio"))]
+impl AsyncProtocolReader {
+    /// Decrypt contents of received packet from buffer.
+    ///
+    /// This function is cancellation safe.
+    ///
+    /// # Arguments
+    ///
+    /// * `buffer` - Asynchronous I/O buffer to pull bytes from.
+    ///
+    /// # Returns
+    ///
+    /// A `Result` containing:
+    ///   * `Ok(Payload)`: A decrypted payload.
+    ///   * `Err(ProtocolError)`: An error that occurred during the read or decryption.
+    pub async fn read_and_decrypt<R>(&mut self, buffer: &mut R) -> Result<Payload, ProtocolError>
+    where
+        R: AsyncRead + Unpin + Send,
+    {
+        // Storing state between async reads to make function cancellation safe.
+        loop {
+            match &mut self.state {
+                DecryptState::ReadingLength {
+                    length_bytes,
+                    bytes_read,
+                } => {
+                    while *bytes_read < 3 {
+                        *bytes_read += buffer.read(&mut length_bytes[*bytes_read..]).await?;
+                    }
+
+                    let packet_bytes_len = self.packet_reader.decypt_len(*length_bytes);
+                    self.state = DecryptState::init_reading_payload(packet_bytes_len);
+                }
+                DecryptState::ReadingPayload {
+                    packet_bytes,
+                    bytes_read,
+                } => {
+                    while *bytes_read < packet_bytes.len() {
+                        *bytes_read += buffer.read(&mut packet_bytes[*bytes_read..]).await?;
+                    }
+
+                    let payload = self.packet_reader.decrypt_payload(packet_bytes, None)?;
+                    self.state = DecryptState::init_reading_length();
+                    return Ok(payload);
+                }
+            }
+        }
+    }
+
+    /// Consume the protocol reader in exchange for the underlying packet decoder.
+    pub fn decoder(self) -> PacketReader {
+        self.packet_reader
+    }
+}
+
+/// Manages an async buffer to automatically encrypt and send contents in packets.
+#[cfg(any(feature = "futures", feature = "tokio"))]
+pub struct AsyncProtocolWriter {
+    packet_writer: PacketWriter,
+}
+
+#[cfg(any(feature = "futures", feature = "tokio"))]
+impl AsyncProtocolWriter {
+    /// Encrypt contents and write packet buffer.
+    ///
+    /// # Arguments
+    ///
+    /// * `buffer` - Asynchronous I/O buffer to write bytes to.
+    ///
+    /// # Returns
+    ///
+    /// A `Result` containing:
+    ///   * `Ok()`: On successful contents encryption and packet send.
+    ///   * `Err(ProtocolError)`: An error that occurred during the encryption or write.
+    pub async fn encrypt_and_write<W>(
+        &mut self,
+        plaintext: &[u8],
+        buffer: &mut W,
+    ) -> Result<(), ProtocolError>
+    where
+        W: AsyncWrite + Unpin + Send,
+    {
+        let write_bytes =
+            self.packet_writer
+                .encrypt_packet(plaintext, None, PacketType::Genuine)?;
+        buffer.write_all(&write_bytes[..]).await?;
+        buffer.flush().await?;
+        Ok(())
+    }
+
+    /// Consume the protocol writer in exchange for the underlying packet encoder.
+    pub fn encoder(self) -> PacketWriter {
+        self.packet_writer
+    }
+}
